pyproject.toml
[project]
name = "fastapi-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.2.0",
    "alembic>=1.16.1",
    "bcrypt>=4.3.0",
    "fastapi>=0.115.12",
    "itsdangerous>=2.2.0",
    "passlib>=1.7.4",
    "pydantic-settings>=2.9.1",
    "python-jose>=3.5.0",
    "requests>=2.32.3",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.2",
]
